+++ LFG/app/src/main/java/com/dat257/team1/LFG/view/loginPage/LoginActivity.java
@@ -23,6 +23,7 @@ import com.google.android.gms.auth.api.signin.GoogleSignIn;
 import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
 import com.google.android.gms.auth.api.signin.GoogleSignInClient;
 import com.google.android.gms.auth.api.signin.GoogleSignInOptions;
+import com.google.android.gms.common.SignInButton;
 import com.google.android.gms.common.api.ApiException;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
@@ -44,7 +45,6 @@ public class LoginActivity extends AppCompatActivity {
     int RC_SIGN_IN = 0;
     private Button loginButton;
     private EditText passwordField, emailField;
-    // private SignInButton googleButton;
     private TextView forgetPassword;
     private FirebaseAuth mAuth;
     private FirebaseUser user;
@@ -77,32 +77,11 @@ public class LoginActivity extends AppCompatActivity {
                 openForgotPassword();
             }
         });
-/*
-        googleButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                googleSignIn();
-            }
-        });
-*/
-        GoogleSignInOptions gso = new GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
-                .requestIdToken(getString(R.string.default_web_client_id))
-                .requestEmail()
-                .build();
 
-        mGoogleSignInClient = GoogleSignIn.getClient(this, gso);
         mAuth = FirebaseAuth.getInstance();
 
     }
 
-
-    /**
-     * Sign in with google
-     */
-    private void googleSignIn() {
-        Intent signInIntent = mGoogleSignInClient.getSignInIntent();
-        startActivityForResult(signInIntent, RC_SIGN_IN);
-    }
 
 
     @Override
