+++ LFG/app/src/main/java/com/dat257/team1/LFG/view/loginPage/WelcomeActivity.java
@@ -5,12 +5,17 @@ import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
 import android.widget.Button;
+import android.widget.ImageButton;
 
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 
 import com.dat257.team1.LFG.R;
 import com.dat257.team1.LFG.view.MenuActivity;
+import com.google.android.gms.auth.api.signin.GoogleSignIn;
+import com.google.android.gms.auth.api.signin.GoogleSignInClient;
+import com.google.android.gms.auth.api.signin.GoogleSignInOptions;
+import com.google.firebase.auth.FirebaseAuth;
 
 /**
  * A view that is responsible for holding the welcome page's components
@@ -22,6 +27,10 @@ public class WelcomeActivity extends AppCompatActivity {
     private Button loginButton;
     private Button registerButton;
     private Button goButton;
+    private FirebaseAuth mAuth;
+    private ImageButton googleButton;
+    int RC_SIGN_IN = 0;
+    GoogleSignInClient mGoogleSignInClient;
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
@@ -30,7 +39,7 @@ public class WelcomeActivity extends AppCompatActivity {
         loginButton = findViewById(R.id.login);
         registerButton = findViewById(R.id.register);
         goButton = findViewById(R.id.go);
-
+        googleButton = findViewById(R.id.googleSignin);
         loginButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
@@ -57,6 +66,27 @@ public class WelcomeActivity extends AppCompatActivity {
             }
         });
 
+        googleButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                googleSignIn();
+            }
+        });
+
+        GoogleSignInOptions gso = new GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
+                .requestIdToken(getString(R.string.default_web_client_id))
+                .requestEmail()
+                .build();
+
+        mGoogleSignInClient = GoogleSignIn.getClient(this, gso);
     }
 
+
+    /**
+     * Sign in with google
+     */
+    private void googleSignIn() {
+        Intent signInIntent = mGoogleSignInClient.getSignInIntent();
+        startActivityForResult(signInIntent, RC_SIGN_IN);
+    }
 }
